requirements.txt
# API + serving
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
python-multipart==0.0.9

# AI + embeddings
openai==1.30.1
sentence-transformers==3.0.1
faiss-cpu==1.8.0
numpy==1.26.4
scikit-learn==1.4.2

# Databases
neo4j==5.20.0
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.30

# Scraping + parsing
aiohttp==3.9.5
beautifulsoup4==4.12.3
pdfplumber==0.11.1
pytesseract==0.3.10

# Observability
structlog==24.1.0
prometheus-fastapi-instrumentator==6.1.0
opentelemetry-sdk==1.24.0

# Dev
pytest==8.2.2
pytest-asyncio==0.23.7
httpx==0.27.0
